requirements.txt
Pillow==11.0.0
numpy==1.26.4
opencv-python==4.11.0.86